deploy/requirements.txt
fastapi
pymongo
bcrypt
python-multipart
pillow